pyproject.toml
[project]
name = "simple-simulation"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "asyncua==1.0.3",
    "commonroad-drivability-checker==2022.1",
    "commonroad-io==2022.1",
    "commonroad-vehicle-models==3.0.2",
    "do-mpc==4.6.0",
    "loguru>=0.7.3",
    "matplotlib>=3.7.5",
    "numpy>=1.24.4",
    "onnx==1.14.0",
    "opencv-python>=4.11.0.86",
    "orjson>=3.10.15",
    "reportlab==4.2.0",
    "shapely==1.8.5.post1",
    "simple-scenario>=0.10.1",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
    "ruff>=0.13.0",
]
